Update quinerelay.pl
This is the finished version. It works perfectly as far as my tests can tell. (ie. running it through the quine) I had to change the perl version a little and made it look ugly, but its the only way I could find to get it to work. Sorry about that >.<